% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_nyc_holidays.R
\name{get_nyc_holidays}
\alias{get_nyc_holidays}
\title{Get NYC holiday dates}
\usage{
get_nyc_holidays(return_data_frame = FALSE)
}
\arguments{
\item{return_data_frame}{A logical vector. Default is FALSE}
}
\value{
a vector of holiday dates. If return_data_frame is TRUE, returns a data frame
}
\description{
\code{get_nyc_holidays} returns a vector of all of the city holiday dates per the Office of Payroll Administration in yyyy-mm-dd format. If return_data_frame is set to TRUE, then returns a data frame with date, holiday description, the date in yyyy-mm-dd format and the day of the week.
}
\details{
The function grabs the dates from the list of holidays pdf file from OPA from the current year.
}
